+++ android/src/main/java/com/pichillilorenzo/flutter_inappwebview/pull_to_refresh/PullToRefreshChannelDelegate.java
@@ -29,8 +29,16 @@ public class PullToRefreshChannelDelegate extends ChannelDelegateImpl {
       case "setEnabled":
         if (pullToRefreshView != null) {
           Boolean enabled = (Boolean) call.argument("enabled");
+          pullToRefreshView.settings.enabled = enabled; // used by InAppWebView.onOverScrolled
           pullToRefreshView.setEnabled(enabled);
           result.success(true);
+        } else {
+          result.success(false);
+        }
+        break;
+      case "isEnabled":
+        if (pullToRefreshView != null) {
+          result.success(pullToRefreshView.isEnabled());
         } else {
           result.success(false);
         }